src/python/api-caller/pyproject.toml
[tool.poetry]
name = "min-web-front"
version = "1.0.0"
description = "Web front project python source."
authors = ["Bruno Araujo <araujo.contato@gmail.com>, Yordan Georgiev <yordan.georgiev@csitea.net>"]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^5.4.1"
Jinja2 = "^3.0.1"
pprintjson = "^1.4.2"
requests = "^2.26.0"
jq = "^1.1.3"
urllib3 = "^1.26.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
